Introduce instance creator asynchronous flow control operator (#3531)
* initial implementation

* rename operator

* add operator to main files and tests

* add constraint validation

* fix operator and implement e2e test

* allow new states for formation assignments

* revert external services mock

* adapt existing unit tests

* make imports

* bump director version in e2e tests

* bump values.yaml

* address linter issues

* fix migration

* add unit tests

* add operator unit tests

* add new consumer type

* add new consumer type to validation

* add new consumer type and adapt test and director

* add unit tests

* remove temporary fixture

* adapt e2e test to new format

* address PR comments

* adapt versions and migration name

* move state checks to a more centralized place

* add godoc

* move one more method

* fix condition

* address comments

* bump director version

* rename migration and bump dependencies

* fix test after merge

+++ components/director/internal/domain/formationassignment/status_service.go
@@ -116,11 +116,13 @@ func (fau *formationAssignmentStatusService) DeleteWithConstraints(ctx context.C
 		return errors.Wrapf(err, "while enforcing constraints for target operation %q and constraint type %q", model.NotificationStatusReturned, model.PreOperation)
 	}
 
-	if err = fau.repo.Delete(ctx, id, tenantID); err != nil {
-		if apperrors.IsUnauthorizedError(err) {
-			return apperrors.NewNotFoundError(resource.FormationAssignment, id)
+	if notificationStatusReport.State == string(model.ReadyAssignmentState) {
+		if err = fau.repo.Delete(ctx, id, tenantID); err != nil {
+			if apperrors.IsUnauthorizedError(err) {
+				return apperrors.NewNotFoundError(resource.FormationAssignment, id)
+			}
+			return errors.Wrapf(err, "while deleting formation assignment with ID: %q", id)
 		}
-		return errors.Wrapf(err, "while deleting formation assignment with ID: %q", id)
 	}
 
 	joinPointDetails.Location = formationconstraint.PostNotificationStatusReturned
